% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Address.R
\name{Address}
\alias{Address}
\title{Address}
\description{
Address Class
}
\examples{
\dontrun{
x <- Address$new("User+tag@example.com")
x
x$local
x$local$valid()
x$host
x$host$valid()
x$valid()
x$fail()
x$munge()
x$normal()

x <- Address$new("User+tag@example.com",
  config = list(munge_string = "*^*^*^*"))
x$munge()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{method}}{(character) one or more URLs}

\item{\code{url}}{(character) one or more URLs}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{method}}{(character) one or more URLs}

\item{\code{url}}{(character) one or more URLs}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Address$new()}}
\item \href{#method-split_local_host}{\code{Address$split_local_host()}}
\item \href{#method-normal}{\code{Address$normal()}}
\item \href{#method-to_s}{\code{Address$to_s()}}
\item \href{#method-munge}{\code{Address$munge()}}
\item \href{#method-canonical}{\code{Address$canonical()}}
\item \href{#method-base}{\code{Address$base()}}
\item \href{#method-reference}{\code{Address$reference()}}
\item \href{#method-valid}{\code{Address$valid()}}
\item \href{#method-fail}{\code{Address$fail()}}
\item \href{#method-clone}{\code{Address$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Address object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Address$new(email_address, config = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email_address}}{(character) an email address}

\item{\code{config}}{(list) list of config options}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-split_local_host"></a>}}
\if{latex}{\out{\hypertarget{method-split_local_host}{}}}
\subsection{Method \code{split_local_host()}}{
split local host
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Address$split_local_host(email)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{(character) email address}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
character string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-normal"></a>}}
\if{latex}{\out{\hypertarget{method-normal}{}}}
\subsection{Method \code{normal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Address$normal()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_s"></a>}}
\if{latex}{\out{\hypertarget{method-to_s}{}}}
\subsection{Method \code{to_s()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Address$to_s()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-munge"></a>}}
\if{latex}{\out{\hypertarget{method-munge}{}}}
\subsection{Method \code{munge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Address$munge()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-canonical"></a>}}
\if{latex}{\out{\hypertarget{method-canonical}{}}}
\subsection{Method \code{canonical()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Address$canonical()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-base"></a>}}
\if{latex}{\out{\hypertarget{method-base}{}}}
\subsection{Method \code{base()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Address$base()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reference"></a>}}
\if{latex}{\out{\hypertarget{method-reference}{}}}
\subsection{Method \code{reference()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Address$reference()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-valid"></a>}}
\if{latex}{\out{\hypertarget{method-valid}{}}}
\subsection{Method \code{valid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Address$valid(options = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fail"></a>}}
\if{latex}{\out{\hypertarget{method-fail}{}}}
\subsection{Method \code{fail()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Address$fail()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Address$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
