% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addressable-package.R
\docType{package}
\name{addressable-package}
\alias{addressable-package}
\alias{addressable}
\title{addressable}
\description{
Email Address Validation
}
\author{
Scott Chamberlain \email{sckott@protonmail.com}
}
\keyword{package}
