% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{Config}
\alias{Config}
\title{Config}
\description{
Config Class
}
\examples{
\dontrun{
x <- Config$new()
x
x$config
x$config$dns_timeout

z <- Config$new(list(dns_timeout = 4))
z
z$config
z$config$dns_timeout
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-all_settings}{\code{Config$all_settings()}}
\item \href{#method-new}{\code{Config$new()}}
\item \href{#method-provider}{\code{Config$provider()}}
\item \href{#method-error_message}{\code{Config$error_message()}}
\item \href{#method-error_messages}{\code{Config$error_messages()}}
\item \href{#method-configure}{\code{Config$configure()}}
\item \href{#method-clone}{\code{Config$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_settings"></a>}}
\if{latex}{\out{\hypertarget{method-all_settings}{}}}
\subsection{Method \code{all_settings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Config$all_settings(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Config$new(config = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-provider"></a>}}
\if{latex}{\out{\hypertarget{method-provider}{}}}
\subsection{Method \code{provider()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Config$provider(name, config = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-error_message"></a>}}
\if{latex}{\out{\hypertarget{method-error_message}{}}}
\subsection{Method \code{error_message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Config$error_message(name, locale = "en")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-error_messages"></a>}}
\if{latex}{\out{\hypertarget{method-error_messages}{}}}
\subsection{Method \code{error_messages()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Config$error_messages(x = list(), locale = "en")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-configure"></a>}}
\if{latex}{\out{\hypertarget{method-configure}{}}}
\subsection{Method \code{configure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Config$configure(settings)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Config$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
