% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exchanger.R
\name{Exchanger}
\alias{Exchanger}
\title{Exchanger}
\description{
Exchanger Class
}
\examples{
\dontrun{
x <- Exchanger$new("gmail.com")
x
x$cached("gmail.com")
x
x$mxers_
x$mxers()
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-cached}{\code{Exchanger$cached()}}
\item \href{#method-new}{\code{Exchanger$new()}}
\item \href{#method-provider}{\code{Exchanger$provider()}}
\item \href{#method-mxers}{\code{Exchanger$mxers()}}
\item \href{#method-mx_ips}{\code{Exchanger$mx_ips()}}
\item \href{#method-matches}{\code{Exchanger$matches()}}
\item \href{#method-in_cidr}{\code{Exchanger$in_cidr()}}
\item \href{#method-clone}{\code{Exchanger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cached"></a>}}
\if{latex}{\out{\hypertarget{method-cached}{}}}
\subsection{Method \code{cached()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exchanger$cached(host, config = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exchanger$new(host, config = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-provider"></a>}}
\if{latex}{\out{\hypertarget{method-provider}{}}}
\subsection{Method \code{provider()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exchanger$provider()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mxers"></a>}}
\if{latex}{\out{\hypertarget{method-mxers}{}}}
\subsection{Method \code{mxers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exchanger$mxers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mx_ips"></a>}}
\if{latex}{\out{\hypertarget{method-mx_ips}{}}}
\subsection{Method \code{mx_ips()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exchanger$mx_ips()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-matches"></a>}}
\if{latex}{\out{\hypertarget{method-matches}{}}}
\subsection{Method \code{matches()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exchanger$matches(rules)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-in_cidr"></a>}}
\if{latex}{\out{\hypertarget{method-in_cidr}{}}}
\subsection{Method \code{in_cidr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exchanger$in_cidr(cidr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exchanger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
