% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Host.R
\name{Host}
\alias{Host}
\title{Host}
\description{
Host Class
}
\examples{
\dontrun{
x <- Host$new("example.com")
x
x$host_name
x$domain_name
x$registration_name
x$tld
x$valid()
# x$munge()
x$error
x$error_message
x$ip_address
x$config$config

x <- Host$new("gmail.com", config = list(dns_lookup = "off"))
x$config$config$dns_lookup
x$valid()

x <- Host$new("gmail.com")
x
x$valid()
x$fail()

x <- Host$new("gm ail.com")
x
x$valid()
x$fail()

x <- Host$new("user1")
x$domain_name
x$dns_name
x$registration_name
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Host$new()}}
\item \href{#method-name}{\code{Host$name()}}
\item \href{#method-canonical}{\code{Host$canonical()}}
\item \href{#method-munge}{\code{Host$munge()}}
\item \href{#method-parse}{\code{Host$parse()}}
\item \href{#method-parse_comment}{\code{Host$parse_comment()}}
\item \href{#method-make_host_name}{\code{Host$make_host_name()}}
\item \href{#method-fully_qualified_domain_name}{\code{Host$fully_qualified_domain_name()}}
\item \href{#method-hosted_service}{\code{Host$hosted_service()}}
\item \href{#method-find_provider}{\code{Host$find_provider()}}
\item \href{#method-set_provider}{\code{Host$set_provider()}}
\item \href{#method-ip}{\code{Host$ip()}}
\item \href{#method-ipv4}{\code{Host$ipv4()}}
\item \href{#method-ipv6}{\code{Host$ipv6()}}
\item \href{#method-matches}{\code{Host$matches()}}
\item \href{#method-registration_name_matches}{\code{Host$registration_name_matches()}}
\item \href{#method-tld_matches}{\code{Host$tld_matches()}}
\item \href{#method-provider_matches}{\code{Host$provider_matches()}}
\item \href{#method-domain_matches}{\code{Host$domain_matches()}}
\item \href{#method-ip_matches}{\code{Host$ip_matches()}}
\item \href{#method-dns_enabled}{\code{Host$dns_enabled()}}
\item \href{#method-dns_a_record}{\code{Host$dns_a_record()}}
\item \href{#method-exchangers}{\code{Host$exchangers()}}
\item \href{#method-valid}{\code{Host$valid()}}
\item \href{#method-fail}{\code{Host$fail()}}
\item \href{#method-valid_dns}{\code{Host$valid_dns()}}
\item \href{#method-valid_mx}{\code{Host$valid_mx()}}
\item \href{#method-valid_format}{\code{Host$valid_format()}}
\item \href{#method-valid_ip}{\code{Host$valid_ip()}}
\item \href{#method-localhost}{\code{Host$localhost()}}
\item \href{#method-clone}{\code{Host$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$new(host_name, config = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-name"></a>}}
\if{latex}{\out{\hypertarget{method-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-canonical"></a>}}
\if{latex}{\out{\hypertarget{method-canonical}{}}}
\subsection{Method \code{canonical()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$canonical()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-munge"></a>}}
\if{latex}{\out{\hypertarget{method-munge}{}}}
\subsection{Method \code{munge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$munge()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parse"></a>}}
\if{latex}{\out{\hypertarget{method-parse}{}}}
\subsection{Method \code{parse()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$parse(host)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parse_comment"></a>}}
\if{latex}{\out{\hypertarget{method-parse_comment}{}}}
\subsection{Method \code{parse_comment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$parse_comment(host)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_host_name"></a>}}
\if{latex}{\out{\hypertarget{method-make_host_name}{}}}
\subsection{Method \code{make_host_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$make_host_name(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fully_qualified_domain_name"></a>}}
\if{latex}{\out{\hypertarget{method-fully_qualified_domain_name}{}}}
\subsection{Method \code{fully_qualified_domain_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$fully_qualified_domain_name(host_part)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hosted_service"></a>}}
\if{latex}{\out{\hypertarget{method-hosted_service}{}}}
\subsection{Method \code{hosted_service()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$hosted_service()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-find_provider"></a>}}
\if{latex}{\out{\hypertarget{method-find_provider}{}}}
\subsection{Method \code{find_provider()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$find_provider()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_provider"></a>}}
\if{latex}{\out{\hypertarget{method-set_provider}{}}}
\subsection{Method \code{set_provider()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$set_provider(name, provider_config = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ip"></a>}}
\if{latex}{\out{\hypertarget{method-ip}{}}}
\subsection{Method \code{ip()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$ip()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ipv4"></a>}}
\if{latex}{\out{\hypertarget{method-ipv4}{}}}
\subsection{Method \code{ipv4()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$ipv4()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ipv6"></a>}}
\if{latex}{\out{\hypertarget{method-ipv6}{}}}
\subsection{Method \code{ipv6()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$ipv6()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-matches"></a>}}
\if{latex}{\out{\hypertarget{method-matches}{}}}
\subsection{Method \code{matches()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$matches(rules)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-registration_name_matches"></a>}}
\if{latex}{\out{\hypertarget{method-registration_name_matches}{}}}
\subsection{Method \code{registration_name_matches()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$registration_name_matches(rule)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tld_matches"></a>}}
\if{latex}{\out{\hypertarget{method-tld_matches}{}}}
\subsection{Method \code{tld_matches()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$tld_matches(rule)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-provider_matches"></a>}}
\if{latex}{\out{\hypertarget{method-provider_matches}{}}}
\subsection{Method \code{provider_matches()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$provider_matches(rule)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-domain_matches"></a>}}
\if{latex}{\out{\hypertarget{method-domain_matches}{}}}
\subsection{Method \code{domain_matches()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$domain_matches(rule)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ip_matches"></a>}}
\if{latex}{\out{\hypertarget{method-ip_matches}{}}}
\subsection{Method \code{ip_matches()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$ip_matches(cidr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dns_enabled"></a>}}
\if{latex}{\out{\hypertarget{method-dns_enabled}{}}}
\subsection{Method \code{dns_enabled()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$dns_enabled()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dns_a_record"></a>}}
\if{latex}{\out{\hypertarget{method-dns_a_record}{}}}
\subsection{Method \code{dns_a_record()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$dns_a_record()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exchangers"></a>}}
\if{latex}{\out{\hypertarget{method-exchangers}{}}}
\subsection{Method \code{exchangers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$exchangers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-valid"></a>}}
\if{latex}{\out{\hypertarget{method-valid}{}}}
\subsection{Method \code{valid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$valid(rules = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fail"></a>}}
\if{latex}{\out{\hypertarget{method-fail}{}}}
\subsection{Method \code{fail()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$fail()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-valid_dns"></a>}}
\if{latex}{\out{\hypertarget{method-valid_dns}{}}}
\subsection{Method \code{valid_dns()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$valid_dns()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-valid_mx"></a>}}
\if{latex}{\out{\hypertarget{method-valid_mx}{}}}
\subsection{Method \code{valid_mx()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$valid_mx()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-valid_format"></a>}}
\if{latex}{\out{\hypertarget{method-valid_format}{}}}
\subsection{Method \code{valid_format()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$valid_format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-valid_ip"></a>}}
\if{latex}{\out{\hypertarget{method-valid_ip}{}}}
\subsection{Method \code{valid_ip()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$valid_ip()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-localhost"></a>}}
\if{latex}{\out{\hypertarget{method-localhost}{}}}
\subsection{Method \code{localhost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$localhost(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
