% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Local.R
\name{Local}
\alias{Local}
\title{Local}
\description{
Local Class
}
\examples{
\dontrun{
host <- Host$new("gmail.com")
x <- Local$new("very.common", list(), host)
x
x$valid()
x$local
x$format()
x$munge()
x$error
x$error_message
x$config$config
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Local$new()}}
\item \href{#method-set_local}{\code{Local$set_local()}}
\item \href{#method-parse}{\code{Local$parse()}}
\item \href{#method-parse_comment}{\code{Local$parse_comment()}}
\item \href{#method-parse_tag}{\code{Local$parse_tag()}}
\item \href{#method-is_ascii}{\code{Local$is_ascii()}}
\item \href{#method-is_unicode}{\code{Local$is_unicode()}}
\item \href{#method-is_redacted}{\code{Local$is_redacted()}}
\item \href{#method-format}{\code{Local$format()}}
\item \href{#method-to_s}{\code{Local$to_s()}}
\item \href{#method-conventional}{\code{Local$conventional()}}
\item \href{#method-canonical}{\code{Local$canonical()}}
\item \href{#method-relax}{\code{Local$relax()}}
\item \href{#method-standard}{\code{Local$standard()}}
\item \href{#method-munge}{\code{Local$munge()}}
\item \href{#method-valid}{\code{Local$valid()}}
\item \href{#method-valid_size}{\code{Local$valid_size()}}
\item \href{#method-valid_size_checks}{\code{Local$valid_size_checks()}}
\item \href{#method-valid_encoding}{\code{Local$valid_encoding()}}
\item \href{#method-is_conventional}{\code{Local$is_conventional()}}
\item \href{#method-is_relaxed}{\code{Local$is_relaxed()}}
\item \href{#method-is_standard}{\code{Local$is_standard()}}
\item \href{#method-clone}{\code{Local$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Local$new(local, config = list(), host = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_local"></a>}}
\if{latex}{\out{\hypertarget{method-set_local}{}}}
\subsection{Method \code{set_local()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Local$set_local(raw)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parse"></a>}}
\if{latex}{\out{\hypertarget{method-parse}{}}}
\subsection{Method \code{parse()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Local$parse(raw)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parse_comment"></a>}}
\if{latex}{\out{\hypertarget{method-parse_comment}{}}}
\subsection{Method \code{parse_comment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Local$parse_comment(raw)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parse_tag"></a>}}
\if{latex}{\out{\hypertarget{method-parse_tag}{}}}
\subsection{Method \code{parse_tag()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Local$parse_tag(raw)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_ascii"></a>}}
\if{latex}{\out{\hypertarget{method-is_ascii}{}}}
\subsection{Method \code{is_ascii()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Local$is_ascii()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_unicode"></a>}}
\if{latex}{\out{\hypertarget{method-is_unicode}{}}}
\subsection{Method \code{is_unicode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Local$is_unicode()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_redacted"></a>}}
\if{latex}{\out{\hypertarget{method-is_redacted}{}}}
\subsection{Method \code{is_redacted()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Local$is_redacted()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Local$format(form = self$config$config[["local_format"]] \%||\% "conventional")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_s"></a>}}
\if{latex}{\out{\hypertarget{method-to_s}{}}}
\subsection{Method \code{to_s()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Local$to_s()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-conventional"></a>}}
\if{latex}{\out{\hypertarget{method-conventional}{}}}
\subsection{Method \code{conventional()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Local$conventional()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-canonical"></a>}}
\if{latex}{\out{\hypertarget{method-canonical}{}}}
\subsection{Method \code{canonical()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Local$canonical()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-relax"></a>}}
\if{latex}{\out{\hypertarget{method-relax}{}}}
\subsection{Method \code{relax()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Local$relax()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-standard"></a>}}
\if{latex}{\out{\hypertarget{method-standard}{}}}
\subsection{Method \code{standard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Local$standard()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-munge"></a>}}
\if{latex}{\out{\hypertarget{method-munge}{}}}
\subsection{Method \code{munge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Local$munge()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-valid"></a>}}
\if{latex}{\out{\hypertarget{method-valid}{}}}
\subsection{Method \code{valid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Local$valid(
  format = self$config$config[["local_format"]] \%||\% "conventional"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-valid_size"></a>}}
\if{latex}{\out{\hypertarget{method-valid_size}{}}}
\subsection{Method \code{valid_size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Local$valid_size()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-valid_size_checks"></a>}}
\if{latex}{\out{\hypertarget{method-valid_size_checks}{}}}
\subsection{Method \code{valid_size_checks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Local$valid_size_checks(range)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-valid_encoding"></a>}}
\if{latex}{\out{\hypertarget{method-valid_encoding}{}}}
\subsection{Method \code{valid_encoding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Local$valid_encoding(
  enc = self$config$config[["local_encoding"]] \%||\% "ascii"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_conventional"></a>}}
\if{latex}{\out{\hypertarget{method-is_conventional}{}}}
\subsection{Method \code{is_conventional()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Local$is_conventional()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_relaxed"></a>}}
\if{latex}{\out{\hypertarget{method-is_relaxed}{}}}
\subsection{Method \code{is_relaxed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Local$is_relaxed()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_standard"></a>}}
\if{latex}{\out{\hypertarget{method-is_standard}{}}}
\subsection{Method \code{is_standard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Local$is_standard()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Local$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
